requirements.txt
pandas
matplotlib
highlight_text
Pillow
urllib3
requests
json
argparse
spotipy
mplsoccer